Using roles to disable pub channels in modal to modify all status

+++ src/Pumukit/NewAdminBundle/Controller/SeriesController.php
@@ -446,10 +446,12 @@ class SeriesController extends AdminController implements NewAdminController
                 foreach($value['channels'] as $channelId => $mustContainsTag){
                     $mustContainsTag = ("true" == $mustContainsTag);
                     $tag = $repoTags->find($channelId);
-                    if ($mustContainsTag && (!($mm->containsTag($tag)))) {
-                        $tagAdded = $tagService->addTagToMultimediaObject($mm, $tag->getId());
-                    }elseif ((!($mustContainsTag)) && $mm->containsTag($tag)) {
-                        $tagAdded = $tagService->removeTagFromMultimediaObject($mm, $tag->getId());
+                    if (!$this->isGranted('ROLE_TAG_DISABLE_' . $tag->getCod())) {
+                        if ($mustContainsTag && (!($mm->containsTag($tag)))) {
+                            $tagAdded = $tagService->addTagToMultimediaObject($mm, $tag->getId());
+                        } elseif ((!($mustContainsTag)) && $mm->containsTag($tag)) {
+                            $tagAdded = $tagService->removeTagFromMultimediaObject($mm, $tag->getId());
+                        }
                     }
                 }
                 $mm->setStatus($value['status']);
